Use servlet 3.0 annotations.
