+++ mdw-hub/src/com/centurylink/mdw/hub/servlet/RootServlet.java
@@ -19,6 +19,7 @@ import java.io.File;
 import java.io.IOException;
 
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -34,6 +35,7 @@ import com.centurylink.mdw.util.file.FileHelper;
  * Also handles the case where index.html itself is customized.
  * TODO: some kind of caching that still provides dynamicism
  */
+@WebServlet(urlPatterns={"/index.html", "/root.js"}, loadOnStartup=1)
 public class RootServlet extends HttpServlet {
 
     private static final String MDW_ADMIN_JS = "<script src=\"js/admin.js\"></script>";
