+++ mdw-hub/src/com/centurylink/mdw/hub/servlet/ServiceApiServlet.java
@@ -19,6 +19,7 @@ import java.io.IOException;
 
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -35,6 +36,7 @@ import io.swagger.util.Yaml;
 /**
  * Scans a service path for Swagger annotations and generates the service spec in JSON or YAML.
  */
+@WebServlet(urlPatterns={"/api/*"}, loadOnStartup=1)
 public class ServiceApiServlet extends HttpServlet {
 
     private static final String YAML_EXT = ".yaml";
