+++ mdw-hub/web/WEB-INF/web.xml
@@ -4,149 +4,7 @@
   xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd" id="mdw-hub" version="3.0">
   <display-name>MDW</display-name>
 
-  <welcome-file-list>
-    <welcome-file>index.html</welcome-file>
-  </welcome-file-list>
-
-  <listener>
-    <listener-class>com.centurylink.mdw.hub.servlet.StartupListener</listener-class>
-  </listener>
-    
-  <filter>
-    <filter-name>AccessFilter</filter-name>
-    <filter-class>com.centurylink.mdw.hub.servlet.AccessFilter</filter-class>
-  </filter>
-  <filter>
-    <filter-name>CustomContentFilter</filter-name>
-    <filter-class>com.centurylink.mdw.hub.servlet.CustomContentFilter</filter-class>
-  </filter>
-
-  <filter-mapping>
-    <filter-name>AccessFilter</filter-name>
-    <url-pattern>/*</url-pattern>
-  </filter-mapping>  
-  <filter-mapping>
-    <filter-name>CustomContentFilter</filter-name>
-    <url-pattern>/*</url-pattern>
-  </filter-mapping>  
-      
-  <servlet>
-    <servlet-name>RootServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.RootServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>LoginServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.LoginServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>ErrorServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.ErrorServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>CustomContentServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.CustomContentServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>AssetContentServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.AssetContentServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>  
-  <servlet>
-    <servlet-name>TestCaseServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.TestCaseServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>ServiceApiServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.ServiceApiServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>RestServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.RestServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>SoapServlet</servlet-name>
-    <servlet-class>com.centurylink.mdw.hub.servlet.SoapServlet</servlet-class>
-    <load-on-startup>1</load-on-startup>
-  </servlet>
-  <servlet>
-    <servlet-name>io.limberest.service.http.RestServlet</servlet-name>
-    <servlet-class>io.limberest.service.http.RestServlet</servlet-class>
-    <!-- limberest should not scan since we keep our own registry -->
-    <init-param>
-      <param-name>io.limberest.scan.packages</param-name>
-      <param-value></param-value>
-    </init-param>
-  </servlet>  
-
-  <servlet-mapping>
-    <servlet-name>RootServlet</servlet-name>
-    <url-pattern>/index.html</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>RootServlet</servlet-name>
-    <url-pattern>/root.js</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>LoginServlet</servlet-name>
-    <url-pattern>/login</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>LoginServlet</servlet-name>
-    <url-pattern>/logout</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>ErrorServlet</servlet-name>
-    <url-pattern>/error</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>CustomContentServlet</servlet-name>
-    <url-pattern>/customContent/*</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>AssetContentServlet</servlet-name>
-    <url-pattern>/asset/*</url-pattern>
-  </servlet-mapping>  
-  <servlet-mapping>
-    <servlet-name>TestCaseServlet</servlet-name>
-    <url-pattern>/testCase/*</url-pattern>
-  </servlet-mapping>  
-  <servlet-mapping>
-    <servlet-name>TestCaseServlet</servlet-name>
-    <url-pattern>/testResult/*</url-pattern>
-  </servlet-mapping>  
-  <servlet-mapping>
-    <servlet-name>ServiceApiServlet</servlet-name>
-    <url-pattern>/api/*</url-pattern>
-  </servlet-mapping>  
-  <servlet-mapping>
-    <servlet-name>RestServlet</servlet-name>
-    <url-pattern>/Services/*</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>RestServlet</servlet-name>
-    <url-pattern>/services/*</url-pattern>
-  </servlet-mapping>
-  <servlet-mapping>
-    <servlet-name>RestServlet</servlet-name>
-    <url-pattern>/REST/*</url-pattern>
-  </servlet-mapping>  
-  <servlet-mapping>
-    <servlet-name>SoapServlet</servlet-name>
-    <url-pattern>/SOAP/*</url-pattern>
-  </servlet-mapping>
-  <!-- override limberest servlet mapping -->
-  <servlet-mapping>
-    <servlet-name>io.limberest.service.http.RestServlet</servlet-name>
-    <url-pattern>/svc/*</url-pattern>
-  </servlet-mapping>  
-
+   <!-- TODO error page in Spring Boot -->
   <error-page>
     <error-code>500</error-code>
     <location>/error</location>
